+++ Cult of the Lamb/Shared/Helpers.cs
@@ -4,6 +4,7 @@ using System.Collections.Generic;
 using System.Diagnostics;
 using System.IO;
 using System.Linq;
+using BepInEx;
 using BepInEx.Logging;
 using MonoMod.Utils;
 using UnityEngine;